requirements.txt
# 核心爬虫库
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0
parsel>=1.8.0

# 异步支持
aiofiles>=23.2.0

# 数据处理
pandas>=2.1.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
pymongo>=4.6.0
redis>=5.0.0

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# 日志
loguru>=0.7.0

# 工具类
tqdm>=4.66.0
fake-useragent>=1.4.0
pillow>=10.1.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 代理支持
requests[socks]>=2.31.0
PySocks>=1.7.1

# 其他工具
click>=8.1.0
rich>=13.7.0
